Clean bundle in CI jobs
Our bundle on the CI keeps growing as we've seen in other projects
(https://github.com/appsignal/appsignal-server/pull/6106). Probably
because when a Ruby gem is upgraded the old version is not removed from
the bundle directory that is later cached by Semaphore.

To ensure our bundle is kept small call `bundle clean` after calling
`bundle install` by adding the `--clean` flag to `bundle install`.

This should speed up our tests by not having to wait for larger than
necessary caches to be downloaded and uploaded.

The `bundle clean` command describes itself as:

> Cleans up unused gems in your bundler directory